+++ WalletWasabi.Fluent/ViewModels/Wallets/Settings/CoinJoinSettingsViewModel.cs
@@ -11,7 +11,7 @@ using WalletWasabi.Fluent.ViewModels.CoinJoinProfiles;
 using WalletWasabi.Fluent.ViewModels.Navigation;
 using WalletWasabi.Wallets;
 
-namespace WalletWasabi.Fluent.ViewModels.Wallets;
+namespace WalletWasabi.Fluent.ViewModels.Wallets.Settings;
 
 [NavigationMetaData(
 	Title = "Coinjoin Settings",
