+++ WalletWasabi.Fluent/ViewModels/Wallets/Settings/VerifyRecoveryWordsViewModel.cs
@@ -14,7 +14,7 @@ using WalletWasabi.Fluent.ViewModels.Navigation;
 using WalletWasabi.Logging;
 using WalletWasabi.Models;
 
-namespace WalletWasabi.Fluent.ViewModels.Wallets;
+namespace WalletWasabi.Fluent.ViewModels.Wallets.Settings;
 
 [NavigationMetaData(Title = "Verify Recovery Words")]
 public partial class VerifyRecoveryWordsViewModel : RoutableViewModel
