+++ WalletWasabi.Fluent/ViewModels/Wallets/Settings/WalletRenameViewModel.cs
@@ -4,7 +4,7 @@ using WalletWasabi.Fluent.Models.Wallets;
 using WalletWasabi.Fluent.Validation;
 using WalletWasabi.Fluent.ViewModels.Dialogs.Base;
 
-namespace WalletWasabi.Fluent.ViewModels.Wallets;
+namespace WalletWasabi.Fluent.ViewModels.Wallets.Settings;
 
 [NavigationMetaData(Title = "Rename Wallet", NavigationTarget = NavigationTarget.CompactDialogScreen)]
 public partial class WalletRenameViewModel : DialogViewModelBase<Unit>
