+++ WalletWasabi.Fluent/Views/Wallets/Settings/CoinJoinSettingsView.axaml
@@ -3,9 +3,9 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             xmlns:wallets="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets"
-             x:DataType="wallets:CoinJoinSettingsViewModel"
-             x:Class="WalletWasabi.Fluent.Views.Wallets.CoinJoinSettingsView"
+             xmlns:vm="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets.Settings"
+             x:DataType="vm:CoinJoinSettingsViewModel"
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Settings.CoinJoinSettingsView"
              x:CompileBindings="True">
 
   <StackPanel Spacing="20">
