+++ WalletWasabi.Fluent/Views/Wallets/Settings/VerifyRecoveryWordsView.axaml
@@ -2,11 +2,11 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:vm="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets"
+             xmlns:vm="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets.Settings"
              x:CompileBindings="True"
              x:DataType="vm:VerifyRecoveryWordsViewModel"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="WalletWasabi.Fluent.Views.Wallets.VerifyRecoveryWordsView">
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Settings.VerifyRecoveryWordsView">
   <ContentArea Title="{Binding Title}"
                NextContent="Verify" EnableNext="True" EnableBack="True" EnableCancel="True">
     <DockPanel VerticalAlignment="Top">
