+++ WalletWasabi.Fluent/Views/Wallets/Settings/WalletBackupView.axaml
@@ -3,10 +3,10 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             xmlns:wallets="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets"
-             x:DataType="wallets:WalletSettingsViewModel"
+             xmlns:vm="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets.Settings"
+             x:DataType="vm:WalletSettingsViewModel"
              x:CompileBindings="True"
-             x:Class="WalletWasabi.Fluent.Views.Wallets.WalletBackupView">
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Settings.WalletBackupView">
 
   <StackPanel Spacing="20" IsVisible="{Binding !IsWatchOnly}">
     <TextBlock Classes="h6" Text="Have you checked your wallet backup?" />
