+++ WalletWasabi.Fluent/Views/Wallets/Settings/WalletBackupView.axaml.cs
@@ -1,7 +1,7 @@
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace WalletWasabi.Fluent.Views.Wallets;
+namespace WalletWasabi.Fluent.Views.Wallets.Settings;
 
 public class WalletBackupView : UserControl
 {
