+++ WalletWasabi.Fluent/Views/Wallets/Settings/WalletRenameView.axaml
@@ -2,17 +2,20 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:wallets="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets"
+             xmlns:vm="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets.Settings"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="WalletWasabi.Fluent.Views.Wallets.WalletRenameView"
-             x:DataType="wallets:WalletRenameViewModel" >
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Settings.WalletRenameView"
+             x:DataType="vm:WalletRenameViewModel" >
   <ContentArea Title="{Binding Title}"
                  Caption="You can modify the name of your wallet"
                  EnableBack="{Binding EnableBack}"
                  EnableCancel="{Binding EnableCancel}"
                  EnableNext="True" NextContent="Save"
                  Height="280" Width="450">
-    <TextBox MaxLength="{StaticResource WalletNameMaxLength}" Text="{Binding NewWalletName}" VerticalAlignment="Top" Watermark="Wallet name">
+    <TextBox MaxLength="{StaticResource WalletNameMaxLength}"
+             Text="{Binding NewWalletName}"
+             VerticalAlignment="Top"
+             Watermark="Wallet name">
       <Interaction.Behaviors>
         <FocusOnAttachedBehavior />
         <TextBoxSelectAllTextBehavior />
