+++ WalletWasabi.Fluent/Views/Wallets/Settings/WalletSettingsView.axaml
@@ -3,11 +3,11 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             xmlns:wallets="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets"
-             xmlns:v="clr-namespace:WalletWasabi.Fluent.Views.Wallets"
-             x:DataType="wallets:WalletSettingsViewModel"
+             xmlns:vm="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets.Settings"
+             xmlns:v="clr-namespace:WalletWasabi.Fluent.Views.Wallets.Settings"
+             x:DataType="vm:WalletSettingsViewModel"
              x:CompileBindings="True"
-             x:Class="WalletWasabi.Fluent.Views.Wallets.WalletSettingsView">
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Settings.WalletSettingsView">
   <ContentArea Title="{Binding Title}"
                EnableNext="True" NextContent="Done"
                EnableCancel="{Binding EnableCancel}"
